+++ src/taipy/gui/renderers/json.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 
 import warnings
 from datetime import date, datetime, time
+from json import JSONEncoder
 from pathlib import Path
 
 from flask.json.provider import DefaultJSONProvider
@@ -20,17 +21,28 @@ from ..icon import Icon
 from ..utils import _date_to_ISO, _MapDict, _TaipyBase
 
 
+def _default(o):
+    if isinstance(o, Icon):
+        return o._to_dict()
+    if isinstance(o, _MapDict):
+        return o._dict
+    if isinstance(o, _TaipyBase):
+        return o.get()
+    if isinstance(o, (datetime, date, time)):
+        return _date_to_ISO(o)
+    if isinstance(o, Path):
+        return str(o)
+    try:
+        raise TypeError(f"Object of type {o.__class__.__name__} is not JSON serializable")
+    except Exception as e:
+        warnings.warn(f"JSONEncoder has thrown {e}")
+        return None
+
+
+class _TaipyJsonEncoder(JSONEncoder):
+    def default(self, o):
+        return _default(o)
+
+
 class _TaipyJsonProvider(DefaultJSONProvider):
-    @staticmethod
-    def default(o):
-        if isinstance(o, Icon):
-            return o._to_dict()
-        elif isinstance(o, _MapDict):
-            return o._dict
-        elif isinstance(o, _TaipyBase):
-            return o.get()
-        elif isinstance(o, (datetime, date, time)):
-            return _date_to_ISO(o)
-        elif isinstance(o, Path):
-            return str(o)
-        warnings.warn(f"Object of type {type(o).__name__} is not JSON serializable")
+    default = staticmethod(_default)
